+++ packages/fluentui/react-northstar/src/themes/teams/types.ts
@@ -71,6 +71,12 @@ import { TreeTitleStylesProps } from '../../components/Tree/TreeTitle';
 import { TableProps } from '../../components/Table/Table';
 import { TableRowProps } from '../../components/Table/TableRow';
 import { TableCellStylesProps } from '../../components/Table/TableCell';
+import { CardStylesProps } from '../../components/Card/Card';
+import { CardPreviewStylesProps } from '../../components/Card/CardPreview';
+import { CardTopControlsStylesProps } from '../../components/Card/CardTopControls';
+import { CardHeaderStylesProps } from '../../components/Card/CardHeader';
+import { CardBodyStylesProps } from '../../components/Card/CardBody';
+import { CardFooterStylesProps } from '../../components/Card/CardFooter';
 
 export type TeamsThemeStylesProps = {
   Accordion?: AccordionProps;
@@ -137,6 +143,12 @@ export type TeamsThemeStylesProps = {
   Table?: TableProps;
   TableRow?: TableRowProps;
   TableCell?: TableCellStylesProps;
+  Card?: CardStylesProps;
+  CardPreview?: CardPreviewStylesProps;
+  CardTopControls?: CardTopControlsStylesProps;
+  CardHeader?: CardHeaderStylesProps;
+  CardBody?: CardBodyStylesProps;
+  CardFooter?: CardFooterStylesProps;
 };
 
 export type TeamsContextualColors = {