+++ packages/mdl2-theme/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log - @uifabric/mdl2-theme
 
-This log was last generated on Fri, 17 Jan 2020 02:32:17 GMT and should not be manually modified.
+This log was last generated on Mon, 23 Mar 2020 12:28:29 GMT and should not be manually modified.
 
+## 0.1.43
+Mon, 23 Mar 2020 12:28:29 GMT
+
+### Patches
+
+- Replace OfficeDev/office-ui-fabric-react with microsoft/fluentui (elcraig@microsoft.com)
 ## 0.1.6
 Fri, 17 Jan 2020 02:32:17 GMT